+++ CHANGELOG.md
@@ -4,6 +4,9 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 Babashka [fs](https://github.com/babashka/fs): file system utility library for Clojure
 
+## Unreleased
+
+- [#181](https://github.com/babashka/fs/issues/181) `delete-tree` now deletes broken symbolic link `root` ([@lread](https://github.com/lread))
 ## 0.5.30 (2025-11-25)
 
 - [#184](https://github.com/babashka/fs/issues/184): new check in `copy-tree` for copying to self too rigid